Port more tests from webpage-spec.
https://github.com/ariya/phantomjs/issues/12439
