+++ test/webpage-spec.js
@@ -1,148 +1,6 @@
 describe("WebPage object", function() {
     var page = new WebPage();
 
-    it("should be able to get any signal handler that are currently set on it", function() {
-        page.onInitialized = undefined;
-        expect(page.onInitialized).toBeUndefined();
-        var onInitialized1 = function() { var x = "x"; };
-        page.onInitialized = onInitialized1;
-        expect(page.onInitialized).toEqual(onInitialized1);
-        var onInitialized2 = function() { var y = "y"; };
-        page.onInitialized = onInitialized2;
-        expect(page.onInitialized).toEqual(onInitialized2);
-        expect(page.onInitialized).toNotEqual(onInitialized1);
-        page.onInitialized = null;
-        // Will only allow setting to a function value, so setting it to `null` returns `undefined`
-        expect(page.onInitialized).toBeUndefined();
-        page.onInitialized = undefined;
-        expect(page.onInitialized).toBeUndefined();
-    });
-
-    it("should be able to get any callback handler that are currently set on it", function() {
-        page.onConfirm = undefined;
-        expect(page.onConfirm).toBeUndefined();
-        var onConfirmFunc1 = function() { return !"x"; };
-        page.onConfirm = onConfirmFunc1;
-        expect(page.onConfirm).toEqual(onConfirmFunc1);
-        var onConfirmFunc2 = function() { return !!"y"; };
-        page.onConfirm = onConfirmFunc2;
-        expect(page.onConfirm).toEqual(onConfirmFunc2);
-        expect(page.onConfirm).toNotEqual(onConfirmFunc1);
-        page.onConfirm = null;
-        // Will only allow setting to a function value, so setting it to `null` returns `undefined`
-        expect(page.onConfirm).toBeUndefined();
-        page.onConfirm = undefined;
-        expect(page.onConfirm).toBeUndefined();
-    });
-
-    it("should be able to get the error signal handler that is currently set on it (currently a special 1-off case)", function() {
-        page.onError = undefined;
-        expect(page.onError).toBeUndefined();
-        var onErrorFunc1 = function() { return !"x"; };
-        page.onError = onErrorFunc1;
-        expect(page.onError).toEqual(onErrorFunc1);
-        var onErrorFunc2 = function() { return !!"y"; };
-        page.onError = onErrorFunc2;
-        expect(page.onError).toEqual(onErrorFunc2);
-        expect(page.onError).toNotEqual(onErrorFunc1);
-        page.onError = null;
-        // Will only allow setting to a function value, so setting it to `null` returns `undefined`
-        expect(page.onError).toBeUndefined();
-        page.onError = undefined;
-        expect(page.onError).toBeUndefined();
-    });
-
-    it("should handle click event", function() {
-        runs(function() {
-            page.evaluate(function() {
-                window.addEventListener('mousedown', function(event) {
-                    window.loggedEvent = window.loggedEvent || {};
-                    window.loggedEvent.mousedown = event;
-                }, false);
-                window.addEventListener('mouseup', function(event) {
-                    window.loggedEvent = window.loggedEvent || {};
-                    window.loggedEvent.mouseup = event;
-                }, false);
-            });
-            page.sendEvent('click', 42, 217);
-        });
-
-        waits(50);
-
-        runs(function() {
-            var event = page.evaluate(function() {
-                return window.loggedEvent;
-            });
-            expect(event.mouseup.clientX).toEqual(42);
-            expect(event.mouseup.clientY).toEqual(217);
-            expect(event.mousedown.clientX).toEqual(42);
-            expect(event.mousedown.clientY).toEqual(217);
-        });
-    });
-
-    it("should handle doubleclick event", function () {
-        runs(function () {
-            page.content = '<input id="doubleClickField" type="text" onclick="document.getElementById(\'doubleClickField\').value=\'clicked\';" ondblclick="document.getElementById(\'doubleClickField\').value=\'doubleclicked\';" oncontextmenu="document.getElementById(\'doubleClickField\').value=\'rightclicked\'; return false;" value="hello"/>';
-            var point = page.evaluate(function () {
-                var el = document.querySelector('input');
-                var rect = el.getBoundingClientRect();
-                return { x: rect.left + Math.floor(rect.width / 2), y: rect.top + (rect.height / 2) };
-            });
-            page.sendEvent('doubleclick', point.x, point.y);
-        });
-
-        waits(50);
-
-        runs(function () {
-            var text = page.evaluate(function () {
-                return document.querySelector('input').value;
-            });
-            expect(text).toEqual("doubleclicked");
-        });
-    });
-
-    it("should handle click with modifier keys", function() {
-        runs(function() {
-            page.evaluate(function() {
-                window.addEventListener('click', function(event) {
-                    window.loggedEvent = window.loggedEvent || {};
-                    window.loggedEvent.click = event;
-                }, false);
-            });
-            page.sendEvent('click', 100, 100, 'left', page.event.modifier.shift);
-        });
-
-        waits(50);
-
-        runs(function() {
-            var event = page.evaluate(function() {
-                return window.loggedEvent.click;
-            });
-            expect(event.shiftKey).toEqual(true);
-        });
-    });
-
-    it("should handle doubleclick with modifier keys", function() {
-        runs(function() {
-            page.evaluate(function() {
-                window.addEventListener('dblclick', function(event) {
-                    window.loggedEvent = window.loggedEvent || {};
-                    window.loggedEvent.dblclick = event;
-                }, false);
-            });
-            page.sendEvent('doubleclick', 100, 100, 'left', page.event.modifier.shift);
-        });
-
-        waits(50);
-
-        runs(function() {
-            var event = page.evaluate(function() {
-                return window.loggedEvent.dblclick;
-            });
-            expect(event.shiftKey).toEqual(true);
-        });
-    });
-
     xit("should handle file uploads", function() {
         runs(function() {
             page.content = '<input type="file" id="file">\n' +
@@ -178,78 +36,6 @@ describe("WebPage object", function() {
             expect(files.length).toEqual(2)
             expect(files.fileNames[0]).toEqual("run-tests.js");
             expect(files.fileNames[1]).toEqual("webpage-spec.js");
-        });
-    });
-
-    it("reports unhandled errors", function() {
-        var lastError = null;
-
-        var page = new require('webpage').create();
-        page.onError = function(message) { lastError = message; };
-
-        runs(function() {
-            page.evaluate(function() {
-                setTimeout(function() { referenceError(); }, 0);
-            });
-        });
-
-        waits(0);
-
-        runs(function() {
-            expect(lastError).toEqual("ReferenceError: Can't find variable: referenceError");
-
-            page.evaluate(function() { referenceError2(); });
-            expect(lastError).toEqual("ReferenceError: Can't find variable: referenceError2");
-
-            page.evaluate(function() { throw "foo"; });
-            expect(lastError).toEqual("foo");
-
-            page.evaluate(function() { throw Error("foo"); });
-            expect(lastError).toEqual("Error: foo");
-        });
-    });
-
-    it("doesn't report handled errors", function() {
-        var hadError    = false;
-        var caughtError = false;
-        var page        = require('webpage').create();
-
-        runs(function() {
-            page.onError = function() { hadError = true; };
-            page.evaluate(function() {
-                caughtError = false;
-
-                try {
-                    referenceError();
-                } catch(e) {
-                    caughtError = true;
-                }
-            });
-
-            expect(hadError).toEqual(false);
-            expect(page.evaluate(function() { return caughtError; })).toEqual(true);
-        });
-    });
-
-    it("reports the sourceURL and line of errors", function() {
-        runs(function() {
-            var e1, e2;
-
-            try {
-                referenceError();
-            } catch (e) {
-                e1 = e;
-            }
-
-            try {
-                referenceError();
-            } catch (e) {
-                e2 = e;
-            }
-
-            expect(e1.sourceURL).toMatch(/webpage-spec.js$/);
-            expect(e1.line).toBeGreaterThan(1);
-            expect(e2.line).toEqual(e1.line + 6);
         });
     });
 
@@ -290,22 +76,6 @@ describe("WebPage object", function() {
             expect(stack[0].file).toEqual("./fixtures/error-helper.js");
             expect(stack[0].line).toEqual(7);
             expect(stack[0]["function"]).toEqual("bar");
-        });
-    });
-
-    it("reports errors that occur in the main context", function() {
-        var error;
-        phantom.onError = function(e) { error = e; };
-
-        runs(function() {
-            setTimeout(function() { zomg(); }, 0);
-        });
-
-        waits(0);
-
-        runs(function() {
-            expect(error.toString()).toEqual("ReferenceError: Can't find variable: zomg");
-            phantom.onError = phantom.defaultErrorHandler;
         });
     });
 
@@ -599,65 +369,6 @@ describe("WebPage object", function() {
         runs(function() {
             expect(handled).toEqual(true);
             server.close();
-        });
-    });
-
-    it("should pass variables to functions properly", function() {
-        var testPrimitiveArgs = function() {
-            var samples = [
-                true,
-                0,
-                "`~!@#$%^&*()_+-=[]\\{}|;':\",./<>?",
-                undefined,
-                null
-            ];
-            for (var i = 0; i < samples.length; i++) {
-                if (samples[i] !== arguments[i]) {
-                    console.log("FAIL");
-                }
-            }
-        };
-
-        var testComplexArgs = function() {
-            var samples = [
-                {a:true, b:0, c:"string"},
-                function() { return true; },
-                [true, 0, "string"],
-                /\d+\w*\//
-            ];
-            for (var i = 0; i < samples.length; i++) {
-                if (typeof samples[i] !== typeof arguments[i] ||
-                    samples[i].toString() !== arguments[i].toString()) {
-                    console.log("FAIL");
-                }
-            }
-        };
-
-        var message;
-        runs(function() {
-            page.onConsoleMessage = function (msg) {
-                message = msg;
-            };
-        });
-
-        waits(0);
-
-        runs(function() {
-            page.evaluate(function() {
-                console.log("PASS");
-            });
-            page.evaluate(testPrimitiveArgs,
-                true,
-                0,
-                "`~!@#$%^&*()_+-=[]\\{}|;':\",./<>?",
-                undefined,
-                null);
-            page.evaluate(testComplexArgs,
-                {a:true, b:0, c:"string"},
-                function() { return true; },
-                [true, 0, "string"],
-                /\d+\w*\//);
-            expect(message).toEqual("PASS");
         });
     });
